requirements.txt
fastapi
uvicorn
pyrogram
tgcrypto
python-dotenv
